공격 매커니즘 수정
Machanism인 Strategy 개선
앉기 및 방패 상태 유지를 위한 코드 개선

+++ Assets/Scripts/UI/Scene/UI_PlayerInfo.cs
@@ -42,7 +42,7 @@ public class UI_PlayerInfo : UI_Scene
         _player = player.GetComponent<Player>();
 
         Image HpBariamge = GetImage((int)Images.HPBar);
-        HpBariamge.fillAmount = _player.m_strStat.m_iHp / _player.m_strStat.m_iMaxHp;
+        HpBariamge.fillAmount = _player.m_strStat.m_iHp / (float)_player.m_strStat.m_iMaxHp;
         Image StaminaBariamge = GetImage((int)Images.STAMINABar);
         StaminaBariamge.fillAmount = _player.m_strStat.m_fStemina / _player.m_strStat.m_fMaxStemina;
